Changed 'web' app
Uses /api/ links instead of /report/. When querying Genotype, get data for finding the experiment you want is the same as with experimentsearch index. Still outputs csv except when the experiment query for genotypes matches multiple experiments. In this case it outputs a .json of the results, with a list for each experiment